requirements.txt
# requirements.txt para el servicio OCR
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python==4.8.1.78
numpy==1.25.2
pydantic==2.5.0
python-dotenv==1.0.0